Files.app: Show dialog when the file is opened in another desktop.
This CL changed the result type of fileBrowserPrivate.taskExecute so that the
result shows how the file is opened.  The result value 'message_sent' means that
the execution message is sent to a proper extension and the extension opens a
file, where we cannot open the new window in the current desktop.  The FileTasks
class in JS refers the result value and opens an alert dialog for the
'message_sent' result.

BUG=336093
TEST=manually

Review URL: https://codereview.chromium.org/131583006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@250078 4ff67af0-8c30-449e-8e8b-ad334ec8d88c