Let MetricsService know about some Android Activities
To tune our crash states more finely, we need to track which Activity is in the
foreground when Chrome is killed. Add piping to let MetricsService know when
Activities are swapped in and out by the user.

The CL makes MetricsService track how many times particular Activities have been
launched, as well as how often Chrome was improperly closed while one of these
Activities was in the foreground (i.e. crashed). These numbers are concatenated
into a histogram, which is then sent to the server.

BUG=321346

Review URL: https://codereview.chromium.org/137623002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@250410 4ff67af0-8c30-449e-8e8b-ad334ec8d88c