Keep track of the list of currently mounted volumes in C++-side VolumeManager.
For the detail, refer the associated bug tracker. In short,
I want to do this for implementing decision procedure for whether
or not to mount zip archives based on originating profiles.

Besides, I believe the new code is more clean--the old code that
constructs the volume list every time FindVolumeById called is skewed.

BUG=343038

Review URL: https://codereview.chromium.org/160483004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@251013 4ff67af0-8c30-449e-8e8b-ad334ec8d88c