Add new prerender origin type for <link rel=next ...> launched prerenders.
By adding rel=next as an origin, we get histogram reporting on many of
our existing histograms.

The one tricky decision is that a prerender can have multiple rel
types (both next and prerender), and that by origin we currently
distinguish samedomain vs crossdomain.

In this CL we treat a prerender with both next and prerender rel types
as rel=prerender, and we do not distinguish cross vs same domain on
rel=next prerenders.

R=tburkard,davidben,isherman@chromium.org
BUG=329963

Review URL: https://codereview.chromium.org/121293002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@251139 4ff67af0-8c30-449e-8e8b-ad334ec8d88c