Fixes memory leak in 149313003.
The leak was caused by the modular testing not forcing the run-and-clear of
clean-up callbacks.  This left a callback owned by the object which increment
the refcount of the object.  These callbacks should be removed and a different
method of ensuring cleanup should be employed.

Ideally we can find a way to get rid of the ref-counting on the object to avoid
this sort of issue in the future, but currently the operation manager lives on
the IO thread and the Operation must work on the FILE thread.  A sequenced task
queue or a worker may be useful.

Previous CL: https://codereview.chromium.org/149313003/

BUG=292956
BUG=337883
BUG=335404

Review URL: https://codereview.chromium.org/170713007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@251984 4ff67af0-8c30-449e-8e8b-ad334ec8d88c