Add support for GetHomeDir for Mac and Windows.
Unifies the Path Service's base::DIR_HOME on Posix and base::DIR_PROFILE on Windows to just be base::DIR_HOME everywhere, and DIR_HOME will now work on Mac.

This removes the AssertIOAllowed check in the Posix implementation because that was only executed in a fallback case that no developer is likely to ever hit. In addition, the we do call this type of function on the UI thread, so either we need to promote the assertion to be at the top of the function or delete it. It seemed unreasonable to disallow using this one key of the path service on the UI thread (the other ones are OK) so I just went for deleting the assertion.

R=benwells@chromium.org

Review URL: https://codereview.chromium.org/159833003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252066 4ff67af0-8c30-449e-8e8b-ad334ec8d88c