Move logic that starts CertLoader from login_utils to user_manager_impl.
The problem with having this in login_utils is that this code doesn't run if
the login is skipped (e.g. on crash recovery), so the certificate loader never
starts in this case.

BUG=341324

TEST=
(1. make sure there are some user certs installed for the primary user)
2.in chrome://settings: 'Add connection', 'Add WiFi', 'Advanced',
set 'EAP method' to EAP-TLS, verify there is a certificate listed under
'User certificates'
3.1. go to chrome://inducebrowsercrashforrealz
3.2. in chrome://flags change any flag and click 'Restart now'
4.1. and 4.2. same as step 2.

Review URL: https://codereview.chromium.org/171333002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252305 4ff67af0-8c30-449e-8e8b-ad334ec8d88c