Add frame color option to packaged app windows.
If app windows are created with a frame color option, that color is used
for the frame. If no color is supplied a native style look is used
instead. This is currently implemented on linux aura and windows only.

This change also removes the --apps-use-native-frame flag on windows.

BUG=339558

Review URL: https://codereview.chromium.org/166443004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252548 4ff67af0-8c30-449e-8e8b-ad334ec8d88c