[Sync] Fix handling of dropped favicons to respect type
Previously, we would locally drop both the image and tracking part of a favicon
if it was unsynced. But, the notion of unsynced is per-type, so this can be
incorrect if only part of a favicon is available. Now, drop logic is per type
as well.

BUG=335493

Review URL: https://codereview.chromium.org/171403006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252702 4ff67af0-8c30-449e-8e8b-ad334ec8d88c