[GCM] Adding a list of accounts present on the client to checkin request
We need to bind device id (profile specific) to accounts running in that
profile. For that reason we are grabbing all of the accounts as tracked
by the ProfileOAuth2TokenService and pass them to Initialize method of
GCMClientImpl. The list of accounts is then passed to CheckinRequest when
the request is happening.

Once the CheckinRequest is done more often than simply on login, we may want
to give a more updated list of accounts than the one present on startup of
Chromium. For now there is no reason to handle that case.

BUG=343203

Review URL: https://codereview.chromium.org/171513004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c