Infinite Suggest - views implementation (See comment 36 on crbug.com/327833 for screenshots)
RenderText: 
Allow layout_text_ to be empty if there is no room to render a single character. 

---------------- 
OmniboxResultView: 
Revises the OmniboxResultView to rely on RenderText for eliding and rendering. 

The match contents and description can not be rendered together as one string. 
See crbug.com/331153 for a possible error case. The relevant TODO is removed. 

----------------------------------------------------------- 
Infinite Suggestion: 
These suggestions show a leading ellipsis to suggest tail completion. Ideally 
they should appear at an offset equivalent to that of omitted text. This is 
accomplished by computing the required width as if the full text is present and 
the actual width required to show the suggestion with leading ellipsis. 

If the window is not wide enough to show the suggestion at the offset of the 
omitted text, the suggestions are shifted towards the start edge such that the 
longest infinite suggestion is visible. The suggestions are never shifted beyond 
the start edge. 

The suggestions are shown aligned to the start edge when directionality of suggestion differs from directionality of UI. 

When the directionality of suggestion differs from directionality of UI, we try 
to show the suggestions such that the leading ellipsis is still vertically 
aligned for all suggestions. 

BUG=327833, 336543, 328699

TBR=msw@chromium.org

Review URL: https://codereview.chromium.org/171153004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@252873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c