sync: Unregister SyncInternalsMessageHandler
Unregisters SyncInternalsMessageHandler as an observer of
ProfileSyncService events during the object's destruction.

Fixes a serious bug that caused a lot of crashes on canary.  If we did
not unregister, then it would be possible to trigger a crash by opening
about:sync, closing it, then doing something that triggers a sync cycle
(eg. navigating to a new page or adding a bookmark).

NOTRY=true
BUG=346244

Review URL: https://codereview.chromium.org/178003005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253009 4ff67af0-8c30-449e-8e8b-ad334ec8d88c