PPAPI/NaCl: Move some nexe loading error reporting to the renderer.
This moves PluginErrorCode to ppb_nacl_private.h so that it can be a parameter to the new ReportLoadError function.

ReportLoadError handles progress events and UMA, as well as the special-case for the infobar that's shown when the client's architecture is missing from the manifest.

Eventually, ReportLoadError will do more (e.g. set properties), and we'll also move other "state transitions" for loading down in to the renderer.

BUG=239656

Review URL: https://codereview.chromium.org/171813011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253148 4ff67af0-8c30-449e-8e8b-ad334ec8d88c