Resubmit: Add a Restore() method to ValueStore and make StorageAPI use it
Resubmit of: https://codereview.chromium.org/165223003/

TBR=mpcomplete@chromium.org
TBR=joaodasilva@chromium.org

---------------------------------
Add a Restore() method to ValueStore and make StorageAPI use it
Take the first steps in preventing the Storage from getting and staying
corrupted. This lets us Restore() a ValueStore when it's corrputed, and tries to
do so in the StorageAPI calls.
This also refactors so that there's a Local and Sync SettingsBackend - this has
the added advantage of meaning we don't decorate the Local stores with Sync
behavior.
BUG=261623

Review URL: https://codereview.chromium.org/176963006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253353 4ff67af0-8c30-449e-8e8b-ad334ec8d88c