Files.app: Let onCopyProgress events report error names instead of obsoleted error codes.
Previously onCopyProgress reports error code numbers to the
FileOperationManager. FileOperationManager tends to refer a name property of the
code, but it failes since the code is just a number.  In addition, the FileError
code is obsoleted and being removed from Files.app.  Thus the CL replaces the
code numbers with error names.

BUG=277949
TEST=manually

Review URL: https://codereview.chromium.org/177183009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253382 4ff67af0-8c30-449e-8e8b-ad334ec8d88c