Add notification mechanism when BrowserActionButton's icon has been updated.
In the browser actions container's chevron menu, a MenuItemView's icon comes from BrowserActionView::GetIconWithBadge() (which comes from BrowserActionButton's icon) when the MenuItemView is created. But, BrowserActionButton's icon may be not yet loaded in that timing because it is read from filesystem in another thread.
To solve this issue, we design an icon state observer in BrowserActionButton, a lifecycle observer in MenuItemView. A helper class whose instance binds to MenuItemView's lifecycle and listens to button's icon state, refresh the host menu item's icon when the button's icon was updated.

R=finnur@chromium.org
BUG=319619

Review URL: https://codereview.chromium.org/148143004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253406 4ff67af0-8c30-449e-8e8b-ad334ec8d88c