Add option to install an ephemeral app to the Windows jump list
This patch adds an entry to the Windows jump list to install an 
ephemeral packaged app. Jump lists are available in Win 7 and 
later. 

BUG=339253
TEST=If an app is launched ephemerally, there should be an
"Install app" item under the "Tasks" category when right-clicking 
on the app's icon in the Windows taskbar. If the app is installed,
the "Install app" item should be absent.

Review URL: https://codereview.chromium.org/176233002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253478 4ff67af0-8c30-449e-8e8b-ad334ec8d88c