Reorganize net/disk_cache into backend specific directories.
We want to unlink the blockfile cache on Android, and it seems that
we'll want more platforms with more varieties of backend in the
future. So let's move the backend into its own subdirectory as we
start to untangle the dependencies from tests/etc... into their
respective backends.

This initial checkin isn't perfect; in particular the tests have a lot
of v2 specific dependencies still.

R=rvargas@chromium.org, cbentzel, rvargas
BUG=331062

Review URL: https://codereview.chromium.org/121643003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253581 4ff67af0-8c30-449e-8e8b-ad334ec8d88c