Reland wallpaper API test
The original CL was reverted here:
codereview.chromium.org/167243002
It failed on linux_chromeos debug build and the reason is we try to decode a 404 response web page
as a jpeg image and it somehow cause a out-of-bounds error.
If the response code is 404, we shouldn't try to decode the response.


BUG=176183

Review URL: https://codereview.chromium.org/167613003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253584 4ff67af0-8c30-449e-8e8b-ad334ec8d88c