sync: null-check SyncedTabDelegate in NotificationServiceSessionsRouter
In every case but one we were null-checking the SyncedTabDelegate returned
from the NotificationDetails, except for one, and that case is crashing on
a null pointer.  There's no guarantee that the tab delegate be there, so
this is a speculative but relatively obvious fix.

BUG=347220
R=maniscalco@chromium.org, zea@chromium.org

Review URL: https://codereview.chromium.org/179833007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253681 4ff67af0-8c30-449e-8e8b-ad334ec8d88c