Moved ParseFileSystemSchemeURL from file_system_url to file_system_util.
This makes it available to renderer code (previously it was only
available to the browser), which will be required to receive file system
objects from Pepper plugins. The static method was not really related to
the FileSystemURL class, it's just used by FileSystemURL's
implementation (among other uses), so it did not belong there anyway.

BUG=345158

Review URL: https://codereview.chromium.org/176263003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@253759 4ff67af0-8c30-449e-8e8b-ad334ec8d88c