Converts MessagePopupCollection into a long-lived object.
This improves the behavior since there were several scenarios where
observer methods would cause re-entrancy into the destructor of
MessagePopupCollection. Since it observes MessageCenter, it knows when
it should show popups and when it shouldn't.

This CL is based on dewittj's crrev.com/114553002 with a fix
for some tests in linux-aura and win because of screen_aura.

BUG=327363
R=dewittj@chromium.org, stevenjb@chromium.org

Review URL: https://codereview.chromium.org/180513002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254101 4ff67af0-8c30-449e-8e8b-ad334ec8d88c