[OSX, OmniTheatre] Handle OriginChip click properly.
Clicking on the OriginChip requires the chip to stay visible until the
determination has been made that the chip was clicked and its callback
is executed.

Previous code disabled the chip in OnSetFocus, so that the following
mouseDown: did not get routed to the OriginChip decoration any more.

This change postpones calling OnSetFocus until immediately before the
ButtonPressed callback is invoked - i.e. mouseDown: has already
traversed all decorations and is not affected by possible visibility
changes caused as a reaction to the click.

For a longer discussion of potential approaches to the problem, see
also https://codereview.chromium.org/165853002/

BUG=338653

Review URL: https://codereview.chromium.org/163913011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254106 4ff67af0-8c30-449e-8e8b-ad334ec8d88c