Converts MessagePopupCollection into a long-lived object (2nd).
This improves the behavior since there were several scenarios where
observer methods would cause re-entrancy into the destructor of
MessagePopupCollection. Since it observes MessageCenter, it knows when
it should show popups and when it shouldn't.

This CL is based on dewittj's crrev.com/114553002 with a fix
for some tests in linux-aura and win because of screen_aura.

This is the reland of r254101 with a fix on tests. The previous
CL broke some tests because the initialization delay reset
their test expectations.

BUG=327363
TBR=dewittj@chromium.org, stevenjb@chromium.org
TEST=message_center_unittests on linux_chromeos

Review URL: https://codereview.chromium.org/184843002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254234 4ff67af0-8c30-449e-8e8b-ad334ec8d88c