mac: Fix to enable tab title font smoothing with CoreAnimation
We now draw the title string directly into the TabView's layer instead of using
a NSTextField subview.
This is so that we can get font smoothing to work on earlier OS
(-canDrawSubviewsIntoLayer was only added in 10.9), and even
when the tab background is a pattern image (when using themes).

BUG=245900
NOTRY=true

Review URL: https://codereview.chromium.org/183123005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254580 4ff67af0-8c30-449e-8e8b-ad334ec8d88c