Fix inconsistencies in left/top of app window bounds for non-native windows
AppWindowFrameView::GetWindowBoundsForClientBounds() and
CustomFrameView::GetWindowBoundsForClientBounds() shift the Y 
coordinate of the window bounds to ensure it is not negative.
This causes BaseNativeAppWindowViews::GetFrameInsets() to
compute the wrong insets.

Therefore we should leave the Y coordinate of the window bounds
negative in AppWindowFrameView::GetWindowBoundsForClientBounds()
and CustomFrameView::GetWindowBoundsForClientBounds().

BUG=346115
TEST=browser_tests (AppWindowBrowserTest.*)

Review URL: https://codereview.chromium.org/176963002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254695 4ff67af0-8c30-449e-8e8b-ad334ec8d88c