Disable bookmark shortcut key and menu item when an extension removes the shortcut key
Expands platform support for removal of the bookmark menu item to
Views, and disables the shortcut key if an extension removes the
shortcut key but does not reassign it. Implements the remaining
functionality of the Remove Bookmark Shortcut Chrome API proposal:
https://docs.google.com/a/chromium.org/document/d/1C2Mle92O9uGlji5y5gGDM5tNJ_tVE1Vb-2xgsZPNDTk

This functionality is unimplemented for Linux/GTK pending the
transition from GTK to Views.

BUG=335655

Review URL: https://codereview.chromium.org/182713004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254701 4ff67af0-8c30-449e-8e8b-ad334ec8d88c