mac: Improve performance of throbber animations.
Don't recreate the icon/throbber view everytime we switch states between
waiting, loading, loaded and crashed states.
We now reuse the same view to show throbbers, favicon and also the toast
animation when the tab crashed.
Enable layer based throbbers with the hope that the earlier performance
regressions are addressed with this.
I was not able to repro the regressions locally, but Chris was able to on
a Mac mini running 10.8.

BUG=19268
NOTRY=true

Review URL: https://codereview.chromium.org/178753003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254708 4ff67af0-8c30-449e-8e8b-ad334ec8d88c