MTP Streaming: Optimize header verification.
Currently, for MTP streaming, we make a special request for the header bytes, to verify that it is indeed a media file.

In the common case of reading from start to finish, this is wasteful, as the originally requested read starts from the 0th byte and includes the header bytes anyways.

This detects that case, and reads the first chunk directly, and then looks at the header bytes within that.

BUG=110119

Review URL: https://codereview.chromium.org/185593008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254818 4ff67af0-8c30-449e-8e8b-ad334ec8d88c