Allow the webstore to re-run the app launcher install flow.
This CL lets the app launcher install flow continue even when
prefs::kAppLauncherHasBeenEnabled is set to true. Only when
EnableAppList is triggered via the button on the webstore.

Note that the webstore page at
https://chrome.google.com/webstore/launcher will not actually show the
button when kAppLauncherHasBeenEnabled is true, but this will allow that
(or an alternative page) to provide a re-enable option.

BUG=256117, 315964, 269541

Review URL: https://codereview.chromium.org/183893016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@254941 4ff67af0-8c30-449e-8e8b-ad334ec8d88c