x11: Simplify some more event-rewrite code.
The WindowTreeHost implementation for X11 currently receives an event from X11
when the keyboard mapping changes in the system. The implementation then calls
into WindowTreeHost::OnKeyboardMappingChanged(), which triggers the
corresponding method in all WindowTreeHostObservers. The EventRewriter is a
WindowTreeHostObserver, and so it can then reset the keycodes it maintains.

But the EventRewriter is also a MessagePumpObserver, which means it already
receives the X11 MappingNotify event, and so the plumbing of the event through
aura is unnecessary.  There is no other consumer of
WindowTreeHostObserver::OnKeyboardMappingChanged(), so remove it from the api.

BUG=343639
R=oshima@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/183893031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@255185 4ff67af0-8c30-449e-8e8b-ad334ec8d88c