aura: Remove client::UserActionClient.
UserActionClient is used to navigate back/forward when the back/forward on a
supported mouse-device is clicked on X11. This can instead be achieved by
installing an event-handler on the browser window.

This does change the behaviour a little on ChromeOS: clicking these navigation
buttons when the cursor is not on top of the browser window will not navigate
after this change. I have confirmed with UX that this is a desirable change.

BUG=319636
R=sky@chromium.org

Review URL: https://codereview.chromium.org/183853037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@255385 4ff67af0-8c30-449e-8e8b-ad334ec8d88c