Remove uses of the Legacy Notification API and tests on Linux Aura.
Blink is removing support for the legacy notification API, and there
are a number of Chromium-sided tests depending on it. Remove the
tests if they don't make sense anymore, or change them to use the
Web Notification API instead.

This patch also enables the Notification browser tests for Aura Linux,
part of interactive_ui_tests, since they run just fine. It are the Desktop
Notification tests which have a dependency on Screen.

BUG=348019
R=atwilson@chromium.org, dewittj@chromium.org, kalman@chromium.org, miket@chromium.org

Review URL: https://codereview.chromium.org/187803005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@255599 4ff67af0-8c30-449e-8e8b-ad334ec8d88c