Support non-SFI mode in NaCl manifest file.
Currently, the flag to enable non-SFI mode is checked just before launching the plugin in the browser process. However, even if the flag is set, we may need to run the plugin in SFI mode.
For example, trivially, if the plugin provides only SFI-mode binary.
To handle such a case, this CL adds entries to NaCl manifest file so that a plugin developer can declare if their plugin provides SFI-mode/non-SFI mode binaries.
In summary, NaCl works in non-SFI mode if;
1) --enable-nacl-nonsfi-mode is set to true, and
2) the plugin provides the binary for non-SFI mode.
So, some checks are moved from the browser to the renderer.

We need similar, but slightly different, a flag for non-SFI mode. Here is the naming rule:
1) enable_nonsfi_mode -> If non-SFI mode is enabled on the browser.
  (In more precise, if it is running on supported platform, and --enable-nacl-nonsfi-mode is set)
2) uses_nonsfi_mode -> If the specified plugin should run on non-SFI mode.
  This happens when the plugin provides the binary for non-SFI mode, and --enable-nacl-nonsfi-mode
  is set.

Note that, in later CLs, we split non-SFI mode nacl_helper from the current nacl_helper with linking it to the newlib.
Then, the uses_nonsfi_mode will become a flag to decide which helper the host should talk to.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3734
TEST=Ran trybot.

Review URL: https://codereview.chromium.org/177113009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@255623 4ff67af0-8c30-449e-8e8b-ad334ec8d88c