Use the default dispatcher where possible for nested message loops.
Notable changes:
 * Add QuitNestedMessageLoop() to client::DispatcherClient, which can be used to
   terminate a nested loop started by RunWithDispatcher().
 * FirstRunDialog is no longer a MessagePumpDispatcher. The default dispatcher
   is used instead, and QuitNestedMessageLoop() is called to terminate the loop
   instead of returning POST_DISPATCH_QUIT_LOOP from the Dispatch() override.

This change was previously committed in r253723 and r254089, but reverted in
r253744 and r254095 because some interactive_ui_tests on windows bots failed on
SimpleMessageBoxViews related change. So this CL aims to reland that CL without
the change in SimpleMessageBoxViews for now.

BUG=342338
R=sky@chromium.org

Review URL: https://codereview.chromium.org/192743007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256093 4ff67af0-8c30-449e-8e8b-ad334ec8d88c