webui: Add updateIndexes method to ArrayDataModel.
Previously there is no way to update multiple items in an ArrayDataModel without
redundant sort.  Files.app uses sort() method for the purpose to avoid redundant
sort, but it does not publish a permute event if the order of items does not
changed. Thus sometimes the file list does not update and it causes flakiness of
tests.

This CL adds updateIndexes method and let the Files.app use it.


BUG=348008
TEST=run ArrayDataModelTest

Review URL: https://codereview.chromium.org/178803005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256193 4ff67af0-8c30-449e-8e8b-ad334ec8d88c